Update examples and add Function.prototype.extend polyfill
