+++ examples/experiment_verification.js
@@ -1,54 +1,53 @@
 var express = require('express');
 var app = express();
-
 var planout = require('../dist/planout.js');
+var extend = require('./polyfills/extend.js');
 
-var Experiment1 = function(args) {
-  var experiment = new planout.Experiment(args);
-  experiment.setup = function() { this.setSalt("Exp1"); }
-  experiment.setup();
-  experiment.assign = function(params, args) {
+var Experiment1 = planout.Experiment.extend({
+  setup: function() {
+    this.setName("Exp1");
+    this.setSalt("Exp1");
+  },
+  assign: function(params, args) {
     params.set('group_size', new planout.Ops.Random.UniformChoice({ 'choices': [1, 10], 'unit': args.userid}));
     params.set('specific_goal', new planout.Ops.Random.BernoulliTrial({'p': 0.8, 'unit': args.userid}));
     if (params.get('specific_goal')) {
       params.set('ratings_per_user_goal', new planout.Ops.Random.UniformChoice({ 'choices': [8, 16, 32, 64], 'unit': args.userid}));
       params.set('ratings_goal', params.get('group_size') * params.get('ratings_per_user_goal'));
     }
-  };
-  experiment.getParamNames = function() { return this.getDefaultParamNames(); }
-  experiment.configureLogger = function() { return; }
-  experiment.log = function(stuff) { return; }
-  experiment.previouslyLogged = function() { return; }
-  return experiment;
-};
+  },
+  getParamNames: function() { return this.getDefaultParamNames(); },
+  configureLogger: function() { return; },
+  log: function(stuff) { return; },
+  previouslyLogged: function() { return; },
+});
 
-var Experiment3 = function(args) {
-  var experiment = new planout.Experiment(args);
-  experiment.setup = function() { this.setSalt("Exp3"); }
-  experiment.setup();
-  experiment.assign = function(params, args) {
+var Experiment3 = planout.Experiment.extend({
+  setup: function() {
+    this.setName("Exp3");
+    this.setSalt("Exp3");
+  },
+  assign: function(params, args) {
     params.set('has_banner', new planout.Ops.Random.BernoulliTrial({ 'p': 0.97, 'unit': args.userid}));
     var cond_probs = [0.5, 0.95];
     params.set('has_feed_stories', new planout.Ops.Random.BernoulliTrial({'p': cond_probs[params.get('has_banner')], 'unit': args.userid}));
     params.set('button_text', new planout.Ops.Random.UniformChoice({'choices': ["I'm a voter", "I'm voting"], 'unit': args.userid}))
-  };
-  experiment.getParamNames = function() { return this.getDefaultParamNames(); }
-  experiment.configureLogger = function() { return; }
-  experiment.log = function(stuff) { return; }
-  experiment.previouslyLogged = function() { return; }
-  return experiment;
-
-}
+  },
+  getParamNames: function() { return this.getDefaultParamNames(); },
+  configureLogger: function() { return; },
+  log: function(stuff) { return; },
+  previouslyLogged: function() { return; },
+});
 
 var experiment1_results = [];
 for (var i = 0; i < 10; i++) {
-  var exp = Experiment1({'userid': i});
+  var exp = new Experiment1({'userid': i});
   experiment1_results.push([exp.get('group_size'), exp.get('ratings_goal')]);
 }
 
 var experiment3_results = [];
 for (var i =0 ; i < 5; i++) {
-  var exp = Experiment3({'userid': i});
+  var exp = new Experiment3({'userid': i});
   experiment3_results.push([exp.get('button_text')]);
 }
