+++ examples/sample_planout_es5.js
@@ -1,42 +1,7 @@
 var express = require('express');
 var app = express();
 var planout = require('../dist/planout.js');
-
-
-/* This is basically a way to easily extend classes in ES5 - source: https://gist.github.com/juandopazo/1367191 */
-Object.getOwnPropertyDescriptors = function getOwnPropertyDescriptors(obj) {
-  var descriptors = {};
-  for (var prop in obj) {
-    if (obj.hasOwnProperty(prop)) {
-      descriptors[prop] = Object.getOwnPropertyDescriptor(obj, prop);
-    }
-  }
-  return descriptors;
-};
- 
-Function.prototype.extend = function extend(proto) {
-    var superclass = this;
-    var constructor;
- 
-    if (!proto.hasOwnProperty('constructor')) {
-      Object.defineProperty(proto, 'constructor', {
-        value: function () {
-            // Default call to superclass as in maxmin classes
-            superclass.apply(this, arguments);
-        },
-        writable: true,
-        configurable: true,
-        enumerable: false
-      });
-    }
-    constructor = proto.constructor;
-    
-    constructor.prototype = Object.create(this.prototype, Object.getOwnPropertyDescriptors(proto));
-    
-    return constructor;
-};
-
-/* End extend helper */
+var extend = require('./polyfills/extend.js');
 
 var DummyExperiment = planout.Experiment.extend({
   setup: function() {
